backend/pyproject.toml
[project]
name = "sc2gameanalizer"
version = "0.1.0"
description = "SC2 Replay Parser and Analyzer"
requires-python = ">=3.13"
dependencies = [
    # Core replay parsing
    "sc2reader>=1.8.0",
    # Web scraping for pro replays
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    # ML similarity matching
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    # Environment configuration
    "python-dotenv>=1.0.0",
    # FastAPI backend
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    # Authentication
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.0.0",
    # Crypto payments
    "web3>=6.0.0",
    "eth-account>=0.9.0",
    "mnemonic>=0.20",
    # PostgreSQL
    "psycopg2-binary>=2.9.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov",
]